+++ CHANGELOG.md
@@ -3,6 +3,10 @@ Tidarr notable changes.
 
 [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) format.
 
+## 📦 1.1.3
+### 🐛 Fixed
+- [Front] Fix download links on processing page
+
 ## 📦 1.1.2
 ### 🚀 Added
 - [API] Add Lidarr queue management support